Move contains, reindex, more constructor and as astype
